requirement.txt
supabase
python-dotenv
logging
fastapi
uvicorn
resquests